Move address functions out of controller
